io.ligolw: moved utility methods from table.io.ligolw
most of them are generic anyway
